src/main/java/com/anirudhology/leetcode/linkedlist/LinkedListCycle.java
package com.anirudhology.leetcode.linkedlist;

import com.anirudhology.leetcode.util.ListNode;

public class LinkedListCycle {

    public boolean hasCycle(ListNode head) {
        // Special case
        if (head == null) {
            return false;
        }
        // Slow and fast pointers
        ListNode slow = head;
        ListNode fast = head;
        // Traverse through the list
        while (fast != null && fast.next != null) {
            slow = slow.next;
            fast = fast.next.next;
            if (slow == fast) {
                return true;
            }
        }
        return false;
    }
}
